ce11_all_REs

Regulatory elements annotated in C. elegans (ce11) according to Serizay et al. 2020, "Tissue-specific profiling reveals distinctive regulatory architectures for ubiquitous, germline and somatic genes", BiorXiv.

data(ce11_all_REs)

Format

GRanges

Source

BiorXiv

References

Serizay et al. 2020, "Tissue-specific profiling reveals distinctive regulatory architectures for ubiquitous, germline and somatic genes", BiorXiv. (DOI)

Examples

data(ce11_all_REs)
table(ce11_all_REs$regulatory_class)

#> 
#>        fwd-promoter        rev-promoter   bidirect-promoter   putative_enhancer 
#>                6616                6480                2181               20271 
#> unassigned_promoter pseudogene_promoter      non-coding_RNA       other_element 
#>                2244                 326                 831                8565 

table(ce11_all_REs$which.tissues)

#> 
#>                               Germline                                Neurons 
#>                                   2184                                   2728 
#>                                 Muscle                                 Hypod. 
#>                                   2962                                   3468 
#>                                Intest.                       Germline_Neurons 
#>                                   2852                                    104 
#>                        Germline_Muscle                        Germline_Hypod. 
#>                                    169                                     64 
#>                       Germline_Intest.                         Neurons_Muscle 
#>                                     52                                   1145 
#>                         Neurons_Hypod.                        Neurons_Intest. 
#>                                    102                                     78 
#>                          Muscle_Hypod.                         Muscle_Intest. 
#>                                    411                                    178 
#>                         Hypod._Intest.                Germline_Neurons_Muscle 
#>                                    812                                     56 
#>                Germline_Neurons_Hypod.               Germline_Neurons_Intest. 
#>                                      7                                     12 
#>                 Germline_Muscle_Hypod.                Germline_Muscle_Intest. 
#>                                     27                                     61 
#>                Germline_Hypod._Intest.                  Neurons_Muscle_Hypod. 
#>                                     24                                    187 
#>                 Neurons_Muscle_Intest.                 Neurons_Hypod._Intest. 
#>                                    208                                     66 
#>                  Muscle_Hypod._Intest.         Germline_Neurons_Muscle_Hypod. 
#>                                    511                                     81 
#>        Germline_Neurons_Muscle_Intest.        Germline_Neurons_Hypod._Intest. 
#>                                     69                                     21 
#>         Germline_Muscle_Hypod._Intest.          Neurons_Muscle_Hypod._Intest. 
#>                                    186                                      0 
#> Germline_Neurons_Muscle_Hypod._Intest.                                   Soma 
#>                                      0                                    956 
#>                                  Ubiq.                           Ubiq.-Biased 
#>                                   2918                                   2506 
#>                           Unclassified                                    Low 
#>                                  20478                                   1831