A function to duplicate bi-directional GRanges

This function splits bi-directional ranges into + and - stranded ranges. It duplicates the ranges which are '*'.

deconvolveBidirectionalPromoters(granges)

Arguments

granges
A stranded GRanges object

Value

GRanges with only '+' and '-' strands. GRanges with '*' strand have been duplicated and split into forward and reverse strands.

Examples

data(ce11_all_REs)
library(GenomicRanges)

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colMeans,
#>     colnames, colSums, dirname, do.call, duplicated, eval, evalq,
#>     Filter, Find, get, grep, grepl, intersect, is.unsorted, lapply,
#>     lengths, Map, mapply, match, mget, order, paste, pmax, pmax.int,
#>     pmin, pmin.int, Position, rank, rbind, Reduce, rowMeans, rownames,
#>     rowSums, sapply, setdiff, sort, table, tapply, union, unique,
#>     unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

proms <- ce11_all_REs[grepl('prom', ce11_all_REs$regulatory_class)]
proms

#> GRanges object with 17847 ranges and 2 metadata columns:
#>                                   seqnames            ranges strand |
#>                                      <Rle>         <IRanges>  <Rle> |
#>    nlp-40,homt-1_chrI_11273_11423     chrI       11273-11423      * |
#>           nlp-40_chrI_15431_15581     chrI       15431-15581      * |
#>                ._chrI_16959_17109     chrI       16959-17109      * |
#>           rcor-1_chrI_24453_24603     chrI       24453-24603      * |
#>           rcor-1_chrI_26903_27053     chrI       26903-27053      - |
#>                               ...      ...               ...    ... .
#>    T23E7.6_chrX_17684894_17685044     chrX 17684894-17685044      - |
#>    T23E7.6_chrX_17686030_17686180     chrX 17686030-17686180      - |
#>      cgt-2_chrX_17694789_17694939     chrX 17694789-17694939      + |
#>          ._chrX_17702756_17702906     chrX 17702756-17702906      * |
#>   H11L12.1_chrX_17711839_17711989     chrX 17711839-17711989      - |
#>                                      regulatory_class which.tissues
#>                                              <factor>      <factor>
#>    nlp-40,homt-1_chrI_11273_11423   bidirect-promoter       Intest.
#>           nlp-40_chrI_15431_15581 unassigned_promoter  Unclassified
#>                ._chrI_16959_17109 unassigned_promoter         Ubiq.
#>           rcor-1_chrI_24453_24603 unassigned_promoter  Unclassified
#>           rcor-1_chrI_26903_27053        rev-promoter         Ubiq.
#>                               ...                 ...           ...
#>    T23E7.6_chrX_17684894_17685044        rev-promoter        Hypod.
#>    T23E7.6_chrX_17686030_17686180        rev-promoter  Unclassified
#>      cgt-2_chrX_17694789_17694939        fwd-promoter       Intest.
#>          ._chrX_17702756_17702906 unassigned_promoter  Unclassified
#>   H11L12.1_chrX_17711839_17711989        rev-promoter      Germline
#>   -------
#>   seqinfo: 6 sequences from an unspecified genome; no seqlengths

table(strand(proms))

#> 
#>    +    -    * 
#> 6616 6480 4751 

proms <- deconvolveBidirectionalPromoters(proms)
proms

#> GRanges object with 22598 ranges and 2 metadata columns:
#>                                   seqnames            ranges strand |
#>                                      <Rle>         <IRanges>  <Rle> |
#>    nlp-40,homt-1_chrI_11273_11423     chrI       11273-11423      + |
#>    nlp-40,homt-1_chrI_11273_11423     chrI       11273-11423      - |
#>           nlp-40_chrI_15431_15581     chrI       15431-15581      + |
#>           nlp-40_chrI_15431_15581     chrI       15431-15581      - |
#>                ._chrI_16959_17109     chrI       16959-17109      + |
#>                               ...      ...               ...    ... .
#>    T23E7.6_chrX_17686030_17686180     chrX 17686030-17686180      - |
#>      cgt-2_chrX_17694789_17694939     chrX 17694789-17694939      + |
#>          ._chrX_17702756_17702906     chrX 17702756-17702906      + |
#>          ._chrX_17702756_17702906     chrX 17702756-17702906      - |
#>   H11L12.1_chrX_17711839_17711989     chrX 17711839-17711989      - |
#>                                      regulatory_class which.tissues
#>                                              <factor>      <factor>
#>    nlp-40,homt-1_chrI_11273_11423   bidirect-promoter       Intest.
#>    nlp-40,homt-1_chrI_11273_11423   bidirect-promoter       Intest.
#>           nlp-40_chrI_15431_15581 unassigned_promoter  Unclassified
#>           nlp-40_chrI_15431_15581 unassigned_promoter  Unclassified
#>                ._chrI_16959_17109 unassigned_promoter         Ubiq.
#>                               ...                 ...           ...
#>    T23E7.6_chrX_17686030_17686180        rev-promoter  Unclassified
#>      cgt-2_chrX_17694789_17694939        fwd-promoter       Intest.
#>          ._chrX_17702756_17702906 unassigned_promoter  Unclassified
#>          ._chrX_17702756_17702906 unassigned_promoter  Unclassified
#>   H11L12.1_chrX_17711839_17711989        rev-promoter      Germline
#>   -------
#>   seqinfo: 6 sequences from an unspecified genome; no seqlengths

table(strand(proms))

#> 
#>     +     -     * 
#> 11367 11231     0