A function to sample GRanges from GRanges

This function takes a given GRanges and returns another GRanges object. The new GRanges has the same number of ranges and the same chromosome, width and strand distributions than the original GRanges.

sampleGRanges(
  x,
  n = NULL,
  width = NULL,
  exclude = FALSE,
  avoid_overlap = FALSE
)

Arguments

x
GRanges object

n
Integer, number of sampled GRanges

width
Integer, width of sampled GRanges

exclude
Boolean, should the original GRanges be excluded?

avoid_overlap
Boolean, should the sampled GRanges not be overlapping?

Value

A GRanges object of length n

Examples

data(ce11_proms)
sampleGRanges(ce11_proms, 100)

#> GRanges object with 100 ranges and 0 metadata columns:
#>         seqnames            ranges strand
#>            <Rle>         <IRanges>  <Rle>
#>     [1]     chrI   1109677-1109827      +
#>     [2]     chrI   1555686-1555836      +
#>     [3]     chrI   2672252-2672402      +
#>     [4]     chrI   5262460-5262610      +
#>     [5]     chrI   6822360-6822510      +
#>     ...      ...               ...    ...
#>    [96]     chrX   8895865-8896015      -
#>    [97]     chrX 10297361-10297511      -
#>    [98]     chrX 11822668-11822818      -
#>    [99]     chrX 15789493-15789643      -
#>   [100]     chrX 16924598-16924748      -
#>   -------
#>   seqinfo: 6 sequences from an unspecified genome; no seqlengths